+++ com/BombingGames/Game/Gameobjects/GameObject.java
@@ -59,14 +59,14 @@ public abstract class GameObject {
      * Updates the logic of the object.
      * @param delta time since last update
      */
-    public abstract void update(int delta);
+    public abstract void update(float delta);
     
     public abstract char getCategory();
     /**
      * Place you static update methods here.
      * @param delta 
      */
-    public static void updateStaticUpdates(int delta){
+    public static void updateStaticUpdates(float delta){
         Sea.staticUpdate(delta);
     }
         
@@ -424,6 +424,4 @@ public abstract class GameObject {
     public boolean shouldBeDestroyed() {
         return destroy;
     }
-    
-    
 }
